Create new tile byte format
Track tiles now have their direction information stored in the high 4
bits of the tile byte, creating a clearer separation between the type of
tile and other persistent data.

This will make the code more managable for adding new tile types and with
the new macros and enums the code is much easier to read and understand.